Create OpertingSystemsFunctions.txt
One of the Operating system function is start and shutdown computer. starting up compute is similar for most of computer or mobile devices, and follow 5 steps such as 
step 1 sending electrical current to circuitry in the computer.
Step 2 the processor chip to reset itself and finds the firmware that contain the start-up instruction.
Step 3 checking many components that related to the process.
Step 4 After successful test, Kernel will load information to RAM.
Step 5 The OS in memory takes control computer and load system configuration information. At last, the user interface appears on the screen